initial consolidation into AppTask
