+++ examples/multi-sensor-app/silabs/src/AppTask.cpp
@@ -16,16 +16,27 @@
  *    limitations under the License.
  */
 
+#if (SL_MATTER_GN_BUILD == 0)
+#include "sl_matter_sensor_config.h"
+#endif // SL_MATTER_GN_BUILD
+
 #include "AppTask.h"
 #include "AppConfig.h"
 #include "AppEvent.h"
 #include "LEDWidget.h"
-#include "SensorManager.h"
+#include <app-common/zap-generated/ids/Attributes.h>
+#include <app-common/zap-generated/ids/Clusters.h>
+#include <app/ConcreteAttributePath.h>
+#include <app/clusters/relative-humidity-measurement-server/CodegenIntegration.h>
+#include <app/clusters/temperature-measurement-server/CodegenIntegration.h>
+#include <app/data-model-provider/AttributeChangeListener.h>
 #include <app/server/Server.h>
 #include <app/util/attribute-storage.h>
 #include <assert.h>
 #include <cmsis_os2.h>
+#include <data-model-providers/codegen/CodegenDataModelProvider.h>
 #include <lib/support/CodeUtils.h>
+#include <lib/support/logging/CHIPLogging.h>
 #include <platform/CHIPDeviceLayer.h>
 #include <platform/silabs/platformAbstraction/SilabsPlatform.h>
 #include <setup_payload/OnboardingCodesUtil.h>
@@ -33,12 +44,22 @@
 #include <setup_payload/SetupPayload.h>
 #include <sl_cmsis_os2_common.h>
 
-#if defined (DISPLAY_ENABLED) && DISPLAY_ENABLED
+#if defined(SL_MATTER_USE_SI70XX_SENSOR) && SL_MATTER_USE_SI70XX_SENSOR
+#include <Si70xxSensor.h>
+#else
+#include <crypto/RandUtils.h>
+#endif // defined(SL_MATTER_USE_SI70XX_SENSOR) && SL_MATTER_USE_SI70XX_SENSOR
+
+#if defined(DISPLAY_ENABLED) && DISPLAY_ENABLED
 #include <SensorsUI.h>
 #endif
 
 using namespace chip;
+using namespace chip::app;
+using namespace chip::app::Clusters;
+using namespace chip::DeviceLayer;
 using namespace chip::DeviceLayer::Silabs;
+using namespace chip::Protocols::InteractionModel;
 
 namespace {
 
@@ -56,6 +77,69 @@ constexpr uint8_t kOccupancyLedId = 1;
 constexpr uint8_t kOccupancyLedId = 0;
 #endif
 
+constexpr EndpointId kOccupancySensorEndpoint   = 1;
+constexpr EndpointId kTemperatureSensorEndpoint = 2;
+constexpr EndpointId kHumiditySensorEndpoint    = 3;
+
+// The > 0 is necessary to avoid causing a type-limits compilation error when the threshold is equal to 0
+// which is specific to the uint8_t data type
+#if SL_MATTER_SENSOR_REPORT_THRESHOLD > 0
+static_assert((SL_MATTER_SENSOR_REPORT_THRESHOLD <= std::numeric_limits<uint8_t>::max()),
+              "The value for SL_MATTER_SENSOR_REPORT_THRESHOLD is greater to the allowed maximum value (255).");
+#endif // SL_MATTER_SENSOR_REPORT_THRESHOLD > 0
+static_assert((SL_MATTER_SENSOR_REPORT_THRESHOLD >= std::numeric_limits<uint8_t>::min()),
+              "The value for SL_MATTER_SENSOR_REPORT_THRESHOLD is inferior to the allowed minimum value (0).");
+constexpr uint8_t kAttributeChangeReportThreshold = (SL_MATTER_SENSOR_REPORT_THRESHOLD);
+
+static_assert((std::numeric_limits<uint32_t>::min() <= SL_MATTER_SENSOR_TIMER_PERIOD_S),
+              "The value for SL_MATTER_SENSOR_TIMER_PERIOD_S is inferior to the allowed minimum value (0).");
+static_assert((SL_MATTER_SENSOR_TIMER_PERIOD_S <= std::numeric_limits<uint32_t>::max()),
+              "The value for SL_MATTER_SENSOR_TIMER_PERIOD_S is greater to the allowed maximum value (4294967295).");
+constexpr chip::System::Clock::Seconds32 kSensorReadPeriod = chip::System::Clock::Seconds32(SL_MATTER_SENSOR_TIMER_PERIOD_S);
+
+uint16_t mLastReportedHumidityValue   = 0;
+int16_t mLastReportedTemperatureValue = 0;
+bool isInitialised                    = false;
+
+class SensorManagerAttributeChangeListener : public DataModel::AttributeChangeListener
+{
+public:
+    void OnAttributeChanged(const ConcreteAttributePath & path, DataModel::AttributeChangeType) override
+    {
+        const bool isOccupancyUpdate = path.mEndpointId == kOccupancySensorEndpoint && path.mClusterId == OccupancySensing::Id &&
+            path.mAttributeId == OccupancySensing::Attributes::Occupancy::Id;
+        const bool isTemperatureUpdate = path.mEndpointId == kTemperatureSensorEndpoint &&
+            path.mClusterId == TemperatureMeasurement::Id &&
+            path.mAttributeId == TemperatureMeasurement::Attributes::MeasuredValue::Id;
+        const bool isHumidityUpdate = path.mEndpointId == kHumiditySensorEndpoint &&
+            path.mClusterId == RelativeHumidityMeasurement::Id &&
+            path.mAttributeId == RelativeHumidityMeasurement::Attributes::MeasuredValue::Id;
+
+        if (isOccupancyUpdate)
+        {
+            OccupancySensingCluster * cluster = OccupancySensing::FindClusterOnEndpoint(path.mEndpointId);
+            VerifyOrReturn(cluster != nullptr);
+
+            AppEvent event                         = {};
+            event.Type                             = AppEvent::kEventType_OccupancyAttributeUpdate;
+            event.Handler                          = AppTask::OccupancyAttributeUpdateEvent;
+            event.OccupancyEvent.occupancyDetected = cluster->IsOccupied();
+            AppTask::GetAppTask().PostEvent(&event);
+            return;
+        }
+
+        if (isTemperatureUpdate || isHumidityUpdate)
+        {
+            AppEvent event = {};
+            event.Type     = AppEvent::kEventType_SensorAttributeUpdate;
+            event.Handler  = AppTask::SensorAttributeUpdateEvent;
+            AppTask::GetAppTask().PostEvent(&event);
+        }
+    }
+};
+
+SensorManagerAttributeChangeListener gAttributeChangeListener;
+
 } // namespace
 
 AppTask AppTask::sAppTask;
@@ -68,10 +152,10 @@ CHIP_ERROR AppTask::AppInit()
     sOccupancyLed.Init(kOccupancyLedId);
     sOccupancyLed.Set(false);
 
-    err = SensorManager::Init();
+    err = InitSensorManager();
     if (err != CHIP_NO_ERROR)
     {
-        ChipLogError(AppServer, "SensorManager::Init failed");
+        ChipLogError(AppServer, "InitSensorManager failed");
         appError(err);
     }
 
@@ -141,7 +225,7 @@ void AppTask::UpdateSensorDisplay(void)
     {
     case kSensorUIEnum::kOccupancySensor:
         GetLCD().SetCustomUI(nullptr);
-        GetLCD().WriteDemoUI(SensorManager::IsOccupancyDetected());
+        GetLCD().WriteDemoUI(IsOccupancyDetected());
         break;
     case kSensorUIEnum::kSensor:
         GetLCD().SetCustomUI(SensorsUI::SensorUI);
@@ -201,7 +285,7 @@ void AppTask::ProccessButtonEvent(AppEvent * event)
     VerifyOrReturn(event != nullptr);
     VerifyOrReturn(event->Type == AppEvent::kEventType_Button);
 
-    SensorManager::ButtonActionTriggered(event);
+    GetAppTask().ButtonActionTriggered(event);
 }
 
 void AppTask::SensorAttributeUpdateEvent(AppEvent * event)
@@ -224,4 +308,210 @@ void AppTask::OccupancyAttributeUpdateEvent(AppEvent * event)
 #ifdef DISPLAY_ENABLED
     sAppTask.UpdateSensorDisplay();
 #endif // DISPLAY_ENABLED
+}
+
+void AppTask::SensorActionTriggered(chip::System::Layer * aLayer, void * aAppState)
+{
+    VerifyOrDieWithMsg(isInitialised, AppServer, "Sensor Action was triggered before the Sensor Manager was initialised!");
+
+    int16_t temperature = 0;
+    uint16_t humidity   = 0;
+
+#if defined(SL_MATTER_USE_SI70XX_SENSOR) && SL_MATTER_USE_SI70XX_SENSOR
+    VerifyOrReturn(SL_STATUS_OK == Si70xxSensor::GetSensorData(humidity, temperature));
+#else
+
+    TemperatureMeasurementCluster * tempCluster = TemperatureMeasurement::FindClusterOnEndpoint(kTemperatureSensorEndpoint);
+    DataModel::Nullable<int16_t> maxTempMeasuredValue =
+        (tempCluster != nullptr) ? tempCluster->GetMaxMeasuredValue() : DataModel::Nullable<int16_t>{};
+
+    RelativeHumidityMeasurementCluster * rhCluster = RelativeHumidityMeasurement::FindClusterOnEndpoint(kHumiditySensorEndpoint);
+    DataModel::Nullable<uint16_t> maxMeasuredHumidityValue =
+        (rhCluster != nullptr) ? rhCluster->GetMaxMeasuredValue() : DataModel::Nullable<uint16_t>{};
+
+    DataModel::Nullable<int16_t> currentTempValue;
+    if (tempCluster != nullptr)
+    {
+        currentTempValue = tempCluster->GetMeasuredValue();
+    }
+    if (currentTempValue.IsNull())
+    {
+        currentTempValue.SetNonNull(Crypto::GetRandU16() % maxTempMeasuredValue.Value());
+    }
+
+    DataModel::Nullable<uint16_t> currentHumidityValue;
+    if (rhCluster != nullptr)
+    {
+        currentHumidityValue = rhCluster->GetMeasuredValue();
+    }
+    if (currentHumidityValue.IsNull())
+    {
+        currentHumidityValue.SetNonNull(Crypto::GetRandU16() % maxMeasuredHumidityValue.Value());
+    }
+
+    if (kAttributeChangeReportThreshold > 3)
+    {
+        humidity    = (currentHumidityValue.Value() + kAttributeChangeReportThreshold / 3) % maxMeasuredHumidityValue.Value();
+        temperature = (currentTempValue.Value() + kAttributeChangeReportThreshold / 3) % maxTempMeasuredValue.Value();
+    }
+    else
+    {
+        humidity    = currentHumidityValue.Value() + 1;
+        temperature = currentTempValue.Value() + 1;
+    }
+
+#endif // defined(SL_MATTER_USE_SI70XX_SENSOR) && SL_MATTER_USE_SI70XX_SENSOR
+
+    if (abs(mLastReportedTemperatureValue - temperature) > kAttributeChangeReportThreshold)
+    {
+        mLastReportedTemperatureValue = temperature;
+    }
+    {
+        DataModel::Nullable<int16_t> tempVal;
+        tempVal.SetNonNull(temperature);
+        VerifyOrReturn(TemperatureMeasurement::SetMeasuredValue(kTemperatureSensorEndpoint, tempVal) == CHIP_NO_ERROR);
+    }
+
+    if (abs(mLastReportedHumidityValue - humidity) > kAttributeChangeReportThreshold)
+    {
+        mLastReportedHumidityValue = humidity;
+    }
+    {
+        DataModel::Nullable<uint16_t> humVal;
+        humVal.SetNonNull(humidity);
+        VerifyOrReturn(RelativeHumidityMeasurement::SetMeasuredValue(kHumiditySensorEndpoint, humVal) == CHIP_NO_ERROR);
+    }
+
+    VerifyOrDieWithMsg(aLayer->StartTimer(kSensorReadPeriod, SensorActionTriggered, nullptr) == CHIP_NO_ERROR, AppServer,
+                       "Failed to start recurring timer!");
+
+    ChipLogDetail(AppServer, "Current temperature value: %d", temperature);
+    ChipLogDetail(AppServer, "Current humidity value: %d", humidity);
+}
+
+CHIP_ERROR AppTask::InitSensorManager()
+{
+#if defined(SL_MATTER_USE_SI70XX_SENSOR) && SL_MATTER_USE_SI70XX_SENSOR
+    VerifyOrDieWithMsg(Si70xxSensor::Init() == SL_STATUS_OK, AppServer, "Failed to initialize the sensor!");
+#endif // defined(SL_MATTER_USE_SI70XX_SENSOR) && SL_MATTER_USE_SI70XX_SENSOR
+
+    DeviceLayer::PlatformMgr().LockChipStack();
+    CodegenDataModelProvider::Instance().RegisterAttributeChangeListener(gAttributeChangeListener);
+    DeviceLayer::PlatformMgr().UnlockChipStack();
+
+    isInitialised = true;
+
+    VerifyOrDieWithMsg(DeviceLayer::PlatformMgr().ScheduleWork([](intptr_t arg) {
+        SensorActionTriggered(&chip::DeviceLayer::SystemLayer(), nullptr);
+    }) == CHIP_NO_ERROR,
+                       AppServer, "Failed to schedule the first SensorCallback!");
+
+    return CHIP_NO_ERROR;
+}
+
+void AppTask::ButtonActionTriggered(AppEvent * aEvent)
+{
+    VerifyOrReturn(aEvent->Type == AppEvent::kEventType_Button);
+
+    TEMPORARY_RETURN_IGNORED DeviceLayer::PlatformMgr().ScheduleWork([](intptr_t arg) {
+        OccupancySensingCluster * cluster = OccupancySensing::FindClusterOnEndpoint(kOccupancySensorEndpoint);
+        VerifyOrReturn(cluster != nullptr);
+        bool state = cluster->IsOccupied();
+        cluster->SetOccupancy(!state);
+    });
+}
+
+Status AppTask::GetMeasuredTemperature(DataModel::Nullable<int16_t> & value)
+{
+    TemperatureMeasurementCluster * cluster = TemperatureMeasurement::FindClusterOnEndpoint(kTemperatureSensorEndpoint);
+    if (cluster == nullptr)
+    {
+        return Status::UnsupportedEndpoint;
+    }
+    value = cluster->GetMeasuredValue();
+    return Status::Success;
+}
+
+Status AppTask::GetMaxMeasuredTemperature(DataModel::Nullable<int16_t> & value)
+{
+    TemperatureMeasurementCluster * cluster = TemperatureMeasurement::FindClusterOnEndpoint(kTemperatureSensorEndpoint);
+    if (cluster == nullptr)
+    {
+        return Status::UnsupportedEndpoint;
+    }
+    value = cluster->GetMaxMeasuredValue();
+    return Status::Success;
+}
+
+Status AppTask::GetMinMeasuredTemperature(DataModel::Nullable<int16_t> & value)
+{
+    TemperatureMeasurementCluster * cluster = TemperatureMeasurement::FindClusterOnEndpoint(kTemperatureSensorEndpoint);
+    if (cluster == nullptr)
+    {
+        return Status::UnsupportedEndpoint;
+    }
+    value = cluster->GetMinMeasuredValue();
+    return Status::Success;
+}
+
+Status AppTask::GetMeasuredHumidity(DataModel::Nullable<uint16_t> & value)
+{
+    RelativeHumidityMeasurementCluster * cluster = RelativeHumidityMeasurement::FindClusterOnEndpoint(kHumiditySensorEndpoint);
+    if (cluster == nullptr)
+    {
+        return Status::UnsupportedEndpoint;
+    }
+    value = cluster->GetMeasuredValue();
+    return Status::Success;
+}
+
+Status AppTask::GetMaxMeasuredHumidity(DataModel::Nullable<uint16_t> & value)
+{
+    RelativeHumidityMeasurementCluster * cluster = RelativeHumidityMeasurement::FindClusterOnEndpoint(kHumiditySensorEndpoint);
+    if (cluster == nullptr)
+    {
+        return Status::UnsupportedEndpoint;
+    }
+    value = cluster->GetMaxMeasuredValue();
+    return Status::Success;
+}
+
+Status AppTask::GetMinMeasuredHumidity(DataModel::Nullable<uint16_t> & value)
+{
+    RelativeHumidityMeasurementCluster * cluster = RelativeHumidityMeasurement::FindClusterOnEndpoint(kHumiditySensorEndpoint);
+    if (cluster == nullptr)
+    {
+        return Status::UnsupportedEndpoint;
+    }
+    value = cluster->GetMinMeasuredValue();
+    return Status::Success;
+}
+
+bool AppTask::IsOccupancyDetected()
+{
+    DeviceLayer::PlatformMgr().LockChipStack();
+    OccupancySensingCluster * cluster = OccupancySensing::FindClusterOnEndpoint(kOccupancySensorEndpoint);
+    bool occupied = (cluster != nullptr) ? cluster->IsOccupied() : false;
+    DeviceLayer::PlatformMgr().UnlockChipStack();
+
+    return occupied;
+}
+
+void MatterPostAttributeChangeCallback(const chip::app::ConcreteAttributePath & attributePath, uint8_t type, uint16_t size,
+                                       uint8_t * value)
+{
+    ClusterId clusterId     = attributePath.mClusterId;
+    AttributeId attributeId = attributePath.mAttributeId;
+    ChipLogProgress(Zcl, "Cluster callback: " ChipLogFormatMEI, ChipLogValueMEI(clusterId));
+
+    switch (clusterId)
+    {
+    case Clusters::Identify::Id:
+        ChipLogProgress(Zcl, "Identify attribute ID: " ChipLogFormatMEI " Type: %u Value: %u, length %u",
+                        ChipLogValueMEI(attributeId), type, *value, size);
+        break;
+
+    default:
+        break;
+    }
 }
